created directory sections and in her creted file sectionApartments.js, fixed main.js file
