+++ src/js/constants/apartmentsConstants/cardsDataApartments.js
